[jenkins] Try to make reporting of aborted builds more reliable.
Sometimes when sh steps fail, they throw exceptions that are identical to the
exceptions that would have been thrown if the sh step were aborted.

This is annoying, because it means there's no reliable way of distinguishing
between test failures and aborted steps.

Except: we can ask sh to not throw exceptions in case of execution failures,
but instead return a status code.

This way we might be able to rely on the fact that exceptions=aborted builds.